+++ netschool/src/com/examw/netschool/dao/PlayRecordDao.java
@@ -21,6 +21,7 @@ import android.util.Log;
  */
 public class PlayRecordDao extends BaseDao {
 	private static final String TAG = "PlayRecordDao";
+	private SQLiteDatabase db;
 	/**
 	 * 构造函数。
 	 * @param context
@@ -37,26 +38,27 @@ public class PlayRecordDao extends BaseDao {
 	public void delete(String [] recordIds){
 		Log.d(TAG, "删除播放记录...");
 		if(recordIds == null || recordIds.length == 0) return;
-		SQLiteDatabase db = null;
-		try {
-			final String delSql = "DELETE FROM  tbl_PlayRecords WHERE id in ('"+ StringUtils.join(recordIds, "','") +"') ";
-			Log.d(TAG, "delete-sql:" + delSql);
-			//初始化
-			db = this.dbHelper.getWritableDatabase();
-			//开启事务
-			db.beginTransaction();
-			//删除数据
-			db.execSQL(delSql);
-			//设置事务成功
-			db.setTransactionSuccessful();
-		}catch (Exception e) {
-			Log.e(TAG, "删除播放记录异常:" + e.getMessage(), e);
-		}finally{
-			if(db != null){
-				//结束事务
-				db.endTransaction();
-				//关闭连接
-				db.close();
+		synchronized(dbHelper){
+			try {
+				final String delSql = "DELETE FROM  tbl_PlayRecords WHERE id in ('"+ StringUtils.join(recordIds, "','") +"') ";
+				Log.d(TAG, "delete-sql:" + delSql);
+				//初始化
+				db = dbHelper.getWritableDatabase();
+				//开启事务
+				db.beginTransaction();
+				//删除数据
+				db.execSQL(delSql);
+				//设置事务成功
+				db.setTransactionSuccessful();
+			}catch (Exception e) {
+				Log.e(TAG, "删除播放记录异常:" + e.getMessage(), e);
+			}finally{
+				if(db != null){
+					//结束事务
+					db.endTransaction();
+					//关闭连接
+					db.close();
+				}
 			}
 		}
 	}
@@ -67,26 +69,27 @@ public class PlayRecordDao extends BaseDao {
 	public void delete(String recordId){
 		Log.d(TAG, "删除播放记录...");
 		if(StringUtils.isBlank(recordId)) return;
-		SQLiteDatabase db = null;
-		try {
-			final String delSql = "DELETE FROM  tbl_PlayRecords WHERE id = ? ";
-			Log.d(TAG, "delete-sql:" + delSql);
-			//初始化
-			db = this.dbHelper.getWritableDatabase();
-			//开启事务
-			db.beginTransaction();
-			//删除数据
-			db.execSQL(delSql, new Object[]{ recordId });
-			//设置事务成功
-			db.setTransactionSuccessful();
-		}catch (Exception e) {
-			Log.e(TAG, "删除播放记录异常:" + e.getMessage(), e);
-		}finally{
-			if(db != null){
-				//结束事务
-				db.endTransaction();
-				//关闭连接
-				db.close();
+		synchronized(dbHelper){
+			try {
+				final String delSql = "DELETE FROM  tbl_PlayRecords WHERE id = ? ";
+				Log.d(TAG, "delete-sql:" + delSql);
+				//初始化
+				db = dbHelper.getWritableDatabase();
+				//开启事务
+				db.beginTransaction();
+				//删除数据
+				db.execSQL(delSql, new Object[]{ recordId });
+				//设置事务成功
+				db.setTransactionSuccessful();
+			}catch (Exception e) {
+				Log.e(TAG, "删除播放记录异常:" + e.getMessage(), e);
+			}finally{
+				if(db != null){
+					//结束事务
+					db.endTransaction();
+					//关闭连接
+					db.close();
+				}
 			}
 		}
 	}
@@ -96,24 +99,26 @@ public class PlayRecordDao extends BaseDao {
 	 */
 	public void deleteByLesson(String lessonId){
 		Log.d(TAG, "删除课程资源["+lessonId+"]下播放记录...");
-		SQLiteDatabase db = null;
-		try {
-			//初始化
-			db = this.dbHelper.getWritableDatabase();
-			//开启事务
-			db.beginTransaction();
-			//删除数据
-			db.execSQL("DELETE FROM  tbl_PlayRecords WHERE lesson_id = ? ", new Object[]{ lessonId });
-			//设置事务成功
-			db.setTransactionSuccessful();
-		}catch (Exception e) {
-			Log.e(TAG, "删除课程资源["+lessonId+"]下播放记录异常:" + e.getMessage(), e);
-		}finally{
-			if(db != null){
-				//结束事务
-				db.endTransaction();
-				//关闭连接
-				db.close();
+		if(StringUtils.isBlank(lessonId)) return;
+		synchronized(dbHelper){
+			try {
+				//初始化
+				db = dbHelper.getWritableDatabase();
+				//开启事务
+				db.beginTransaction();
+				//删除数据
+				db.execSQL("DELETE FROM  tbl_PlayRecords WHERE lesson_id = ? ", new Object[]{ lessonId });
+				//设置事务成功
+				db.setTransactionSuccessful();
+			}catch (Exception e) {
+				Log.e(TAG, "删除课程资源["+lessonId+"]下播放记录异常:" + e.getMessage(), e);
+			}finally{
+				if(db != null){
+					//结束事务
+					db.endTransaction();
+					//关闭连接
+					db.close();
+				}
 			}
 		}
 	}
@@ -124,28 +129,29 @@ public class PlayRecordDao extends BaseDao {
 	public void add(PlayRecord  data){
 		Log.d(TAG, "新增播放记录...");
 		if(data == null) return;
-		SQLiteDatabase db = null;
-		try {
-			//初始化
-			db = this.dbHelper.getWritableDatabase();
-			//开启事务
-			db.beginTransaction();
-			//创建记录ID。
-			data.setId(UUID.randomUUID().toString());
-			//新增记录
-			db.execSQL("INSERT INTO tbl_PlayRecords(id,lesson_id,playTime) values (?,?,?)", new Object[]{
-					data.getId(), data.getLessonId(), data.getPlayTime() == null ? 0 : data.getPlayTime().intValue()
-			});
-			//设置事务成功
-			db.setTransactionSuccessful();
-		}catch (Exception e) {
-			Log.e(TAG, "新增播放记录异常:" + e.getMessage(), e);
-		}finally{
-			if(db != null){
-				//结束事务
-				db.endTransaction();
-				//关闭连接
-				db.close();
+		synchronized(dbHelper){
+			try {
+				//初始化
+				db = dbHelper.getWritableDatabase();
+				//开启事务
+				db.beginTransaction();
+				//创建记录ID。
+				data.setId(UUID.randomUUID().toString());
+				//新增记录
+				db.execSQL("INSERT INTO tbl_PlayRecords(id,lesson_id,playTime) values (?,?,?)", new Object[]{
+						data.getId(), data.getLessonId(), data.getPlayTime() == null ? 0 : data.getPlayTime().intValue()
+				});
+				//设置事务成功
+				db.setTransactionSuccessful();
+			}catch (Exception e) {
+				Log.e(TAG, "新增播放记录异常:" + e.getMessage(), e);
+			}finally{
+				if(db != null){
+					//结束事务
+					db.endTransaction();
+					//关闭连接
+					db.close();
+				}
 			}
 		}
 	}
@@ -157,24 +163,25 @@ public class PlayRecordDao extends BaseDao {
 	public void updatePlayTime(String recordId, Integer playTime){
 		Log.d(TAG, "更新["+recordId+"]播放时间..." + playTime);
 		if(StringUtils.isBlank(recordId) || playTime == null) return;
-		SQLiteDatabase db = null;
-		try {
-			//初始化
-			db = this.dbHelper.getWritableDatabase();
-			//开启事务
-			db.beginTransaction();
-			//更新记录
-			db.execSQL("UPDATE tbl_PlayRecords SET playTime = ?  WHERE id = ? ", new Object[]{ playTime, recordId});
-			//设置事务成功
-			db.setTransactionSuccessful();
-		}catch (Exception e) {
-			Log.e(TAG, "新增播放记录异常:" + e.getMessage(), e);
-		}finally{
-			if(db != null){
-				//结束事务
-				db.endTransaction();
-				//关闭连接
-				db.close();
+		synchronized(dbHelper){
+			try {
+				//初始化
+				db = dbHelper.getWritableDatabase();
+				//开启事务
+				db.beginTransaction();
+				//更新记录
+				db.execSQL("UPDATE tbl_PlayRecords SET playTime = ?  WHERE id = ? ", new Object[]{ playTime, recordId});
+				//设置事务成功
+				db.setTransactionSuccessful();
+			}catch (Exception e) {
+				Log.e(TAG, "新增播放记录异常:" + e.getMessage(), e);
+			}finally{
+				if(db != null){
+					//结束事务
+					db.endTransaction();
+					//关闭连接
+					db.close();
+				}
 			}
 		}
 	}
@@ -185,17 +192,13 @@ public class PlayRecordDao extends BaseDao {
 	public List<PlayRecord> loadPlayRecords(){
 		Log.d(TAG, "加载播放记录...");
 		final List<PlayRecord> records = new ArrayList<PlayRecord>();
-		SQLiteDatabase db = null;
 		try {
-			//if(pageIndex == null) pageIndex = 0;
-			//Integer start = (pageIndex - 1) * Constant.PAGE_SIZE;
-			//sql
 			final String query = "SELECT a.id,a.lesson_id,b.name,a.playTime,a.createTime From tbl_PlayRecords a "
 					+ " INNER JOIN tbl_Lessones b ON b.id = a.lesson_id ORDER BY a.createTime DESC";
 			//
 			Log.d(TAG, query);
 			//初始化
-			db = this.dbHelper.getReadableDatabase();
+			db = dbHelper.getReadableDatabase();
 			//查询数据
 			final Cursor cursor =  db.rawQuery(query, null);
 			//循环赋值
@@ -232,16 +235,15 @@ public class PlayRecordDao extends BaseDao {
 	 */
 	public PlayRecord getPlayRecord(String recordId){
 		Log.d(TAG, "加载播放记录..." + recordId);
-		if(StringUtils.isBlank(recordId)) return null;
 		PlayRecord data = null;
-		SQLiteDatabase db = null;
+		if(StringUtils.isBlank(recordId)) return data;
 		try{
 			//sql
 			final String query = "SELECT a.id,a.lesson_id,b.name,a.playTime,a.createTime From tbl_PlayRecords a "
 					+ " INNER JOIN tbl_Lessones b ON b.id = a.lesson_id WHERE a.id = ?";
 			Log.d(TAG, query);
 			//初始化
-			db = this.dbHelper.getReadableDatabase();
+			db = dbHelper.getReadableDatabase();
 			//查询数据
 			final Cursor cursor =  db.rawQuery(query, new String[]{ StringUtils.trimToEmpty(recordId) });
 			//循环赋值
